+++ examples/multiprocessing/pool_initializer.py
@@ -1,9 +1,5 @@
-import logging
-
-from lithops import multiprocessing as mp
-
-
 # import multiprocessing as mp
+from lithops import multiprocessing as mp
 
 
 def work(num):
@@ -17,10 +13,7 @@ def initializer_function(arg1, arg2):
     param2 = arg2
 
 
-# logging.basicConfig(level=logging.DEBUG)
-# logging.getLogger(mp.__name__).setLevel(logging.DEBUG)
-
-
-with mp.Pool(initializer=initializer_function, initargs=('important global arg', 123456)) as p:
-    res = p.map(work, [0] * 3)
-    print(res)
+if __name__ == '__main__':
+    with mp.Pool(initializer=initializer_function, initargs=('important global arg', 123456)) as p:
+        res = p.map(work, [0] * 3)
+        print(res)